// src/types/curriculum.ts
export enum CurriculumItemType {
  Lecture = 'lecture',
  Article = 'article',
  Resource = 'resource',
  Quiz = 'quiz',
  Assignment = 'assignment',
}

export enum EnrollmentStatus {
  Active = 'active',
  Completed = 'completed',
  Dropped = 'dropped',
}

export enum Language {
  English = 'english',
  French = 'french',
  Spanish = 'spanish',
  Yoruba = 'yoruba',
  Igbo = 'igbo',
  Hausa = 'hausa',
}